+++ src/components/bioModule/index.js
