+++ src/components/bioModule/index.module.scss
