+++ src/components/faqCTA/index.module.scss
