+++ src/components/volunteerCTA/index.js
@@ -2,7 +2,7 @@ import React from "react"
 import Button from 'react-bootstrap/Button'
 
 // Style Imports
-import styles from './volunteerCTA.module.scss'
+import styles from './index.module.scss'
 
 const VolunteerCTA = () => {
   return (
